Setup example project, extract `IOChannels`
