+++ .github/workflows/ci.yml
@@ -107,11 +107,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p target .js/target .jvm/target .native/target tests/jvm/target core/target tests/native/target project/target
+        run: mkdir -p target .js/target .jvm/target .native/target tests/jvm/target core/target example/target tests/native/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar target .js/target .jvm/target .native/target tests/jvm/target core/target tests/native/target project/target
+        run: tar cf targets.tar target .js/target .jvm/target .native/target tests/jvm/target core/target example/target tests/native/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
